Fix most remaining icc remarks
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3419 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ framework/modules/profile/src/ProfileModule.cc
@@ -108,7 +108,7 @@ void ProfileModule::terminate()
     int nPoints = m_nEvents / m_step;
     TGraph* graph = new TGraph(nPoints);
     for (int i = 0; i < nPoints; i++) {
-      graph->SetPoint(i, m_eventInfo[i].time, m_eventInfo[i].mem);
+      graph->SetPoint(i, m_eventInfo[i].time, (double)m_eventInfo[i].mem);
     }
     graph->SetMarkerStyle(kMultiply);
     graph->Draw("ALP");
@@ -120,5 +120,3 @@ void ProfileModule::terminate()
     saveDir->cd();
   }
 }
-
-
